requirements.txt
blessings==1.7
bpython==0.18
certifi==2019.9.11
chardet==3.0.4
curtsies==0.3.0
greenlet==0.4.15
idna==2.8
Pygments==2.4.2
pyserial==3.4
requests==2.22.0
six==1.13.0
typing==3.7.4.1
urllib3==1.25.7
wcwidth==0.1.7
